requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
psycopg2-binary==2.8.6
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.6
